+++ lib/metric_fu/reporting/graphs/engines/bluff.rb
@@ -110,4 +110,18 @@ module MetricFu
       File.open(File.join(MetricFu.output_directory, 'rails_best_practices.js'), 'w') {|f| f << content }
     end
   end
+
+  class CaneBluffGrapher < CaneGrapher
+    def graph!
+      content = <<-EOS
+        #{BLUFF_DEFAULT_OPTIONS}
+        g.title = 'Cane: code quality threshold violations';
+        g.data('cane', [#{@cane_violations.join(',')}]);
+        g.labels = #{@labels.to_json};
+        g.draw();
+      EOS
+      File.open(File.join(MetricFu.output_directory, 'cane.js'), 'w') {|f| f << content }
+    end
+  end
+
 end