+++ HISTORY.md
@@ -1,7 +1,12 @@
 === master ===
 
+== MetricFu 3.0.2 / 2013-03-02
+
+* Adding cane metrics (Sathish, pull request #49)
+  * Not yet included in hotspots
+
 == MetricFu 3.0.1 / 2013-03-01
-    
+
 * Fixed typo in Flay generator (Sathish, pull request #47)
 
 === MetricFu 3.0.0 / 2013-02-07